+++ src/types/Client.d.ts
@@ -23,13 +23,15 @@ export interface ClientConfig {
   headers?: { [key: string]: string | number }
   fetch?: typeof fetch
   http2SessionIdleTime?: number
-  notifyAboutNewVersion?: boolean
+  checkNewVersion?: boolean
 }
 
 export interface QueryOptions
   extends Partial<
     Pick<ClientConfig, 'secret' | 'queryTimeout' | 'fetch' | 'observer'>
-  > {}
+  > {
+    signal?: AbortSignal
+  }
 
 type StreamFn<T> = (
   expr: Expr,