+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 4.4.1
+
+- Make `signal` property of `client.query` optional
+
+## 4.4.0
+
+- Improve types (errors.pargeJSON, errors.TooManyRequests, `signal` for client.query options)
+- Fix false positive cases for ClientClosed error
+- Option to disable checking driver version (new Client({checkNewVersion: false}))
+- Fix PageHelper honors size with map/filter
+- Fix module issue for `Vitessedge`
+
 ## 4.3.0
 
 - Implements Client#close method, allows Infinity for `http2SessionIdleTime` parameter